CSE Machine: implement variable declaration check (#1517)
* add validation function to CSE machine

* cleanup: moved all validation functions to src/transpiler

* add test for undefined unreached variables

* checkProgramForUndefinedVariables now considers prelude

* checkProgramForUndefinedVariables now checks for current environment

* reformatting

this is why you don't do --no-verify when pushing

* errors from the validator are now handled normally

* update snapshots
